+++ app/src/app/SnowmanPublicState.ts
@@ -1,0 +1,6 @@
+import { ViewIDs } from 'types/ViewIDs';
+
+export interface SnowmanPublicState {
+  existsActiveRequest: boolean;
+  activeApp: ViewIDs;
+}
