+++ app/src/components/GenericSubApp/GenericSubAppProps.ts
@@ -1,21 +1,14 @@
+import { SnowmanPublicState } from 'app/SnowmanPublicState';
 import { FC } from 'react';
 import { Store } from 'redux';
 import { SnowmanAction } from 'store/messages';
 import { ViewIDs } from 'types/ViewIDs';
 import { SideMenuProps } from 'types/ViewMetaInformation';
 
-export interface GenericSubAppOwnProps {
+export interface GenericSubAppProps extends SnowmanPublicState {
   appTitle: string;
   appId: ViewIDs;
   sideMenu?: FC<SideMenuProps>;
   children?: JSX.Element | JSX.Element[];
   createSubAppStore(): Store<unknown, SnowmanAction>;
 }
-
-export interface GenericSubAppStateProps {
-  isSubAppActive: boolean;
-  existsActiveRequest: boolean;
-}
-
-export type GenericSubAppProps = GenericSubAppOwnProps &
-  GenericSubAppStateProps;
