+++ app/src/pages/BenchmarkPage/BenchmarkApp.Container.tsx
@@ -1,0 +1,26 @@
+import BenchmarkAppView from 'pages/BenchmarkPage/BenchmarkApp.View';
+import { BenchmarkAppDispatchProps } from 'pages/BenchmarkPage/BenchmarkAppProps';
+import { BenchmarkAppDispatch } from 'pages/BenchmarkPage/store/BenchmarkAppModel';
+import {
+  getAlgorithms,
+  getDatasets,
+  getExperiments,
+} from 'pages/BenchmarkPage/store/BenchmarkAppThunkActions';
+import { connect } from 'react-redux';
+
+const mapDispatchToProps = (
+  dispatch: BenchmarkAppDispatch
+): BenchmarkAppDispatchProps => ({
+  loadInitialState: () => {
+    dispatch(getAlgorithms()).then();
+    dispatch(getDatasets()).then();
+    dispatch(getExperiments()).then();
+  },
+});
+
+const BenchmarkAppContainer = connect(
+  null,
+  mapDispatchToProps
+)(BenchmarkAppView);
+
+export default BenchmarkAppContainer;
