+++ app/src/pages/BenchmarkPage/BenchmarkApp.View.tsx
@@ -10,168 +10,156 @@ import {
   IonRow,
   IonText,
 } from '@ionic/react';
-import GenericSubApp from 'components/GenericSubApp/GenericSubApp';
 import { barChart, calculator, colorFilter, pauseCircle } from 'ionicons/icons';
 import { BenchmarkAppProps } from 'pages/BenchmarkPage/BenchmarkAppProps';
-import BenchmarkSelector from 'pages/BenchmarkPage/components/BenchmarkSelector/BenchmarkSelector';
-import { createBenchmarkAppStore } from 'pages/BenchmarkPage/store/BenchmarkAppStoreFactory';
 import React, { useEffect } from 'react';
-import { ViewIDs } from 'types/ViewIDs';
 
 const BenchmarkAppView = ({
   loadInitialState,
 }: BenchmarkAppProps): JSX.Element => {
   useEffect(loadInitialState, [loadInitialState]);
   return (
-    <GenericSubApp
-      appId={ViewIDs.DASHBOARD}
-      appTitle={'Benchmark Dashboard'}
-      createSubAppStore={createBenchmarkAppStore}
-      sideMenu={BenchmarkSelector}
-    >
-      <IonGrid>
-        <IonRow>
-          <IonCol size="4" sizeXl="3">
-            <IonCard>
-              <IonCardHeader>
-                <IonCardTitle>
-                  N-Intersection Viewer
-                  <span>
-                    <IonIcon
-                      className="ion-float-right"
-                      icon={colorFilter}
-                      size="large"
-                      color="primarydark"
-                    />
-                  </span>
-                </IonCardTitle>
-              </IonCardHeader>
-              <IonCardContent>
+    <IonGrid>
+      <IonRow>
+        <IonCol size="4" sizeXl="3">
+          <IonCard>
+            <IonCardHeader>
+              <IonCardTitle>
+                N-Intersection Viewer
+                <span>
+                  <IonIcon
+                    className="ion-float-right"
+                    icon={colorFilter}
+                    size="large"
+                    color="primarydark"
+                  />
+                </span>
+              </IonCardTitle>
+            </IonCardHeader>
+            <IonCardContent>
+              <p>
+                Use a visual representation to investigate which pairs were
+                detected as duplicates within each experiment. Selecting one
+                experiment as a gold standard is optional.
+              </p>
+            </IonCardContent>
+            <IonCardContent>
+              <IonText color="dark">
                 <p>
-                  Use a visual representation to investigate which pairs were
-                  detected as duplicates within each experiment. Selecting one
-                  experiment as a gold standard is optional.
+                  <b>Requires:</b> at least two experiments from the same
+                  dataset
                 </p>
-              </IonCardContent>
-              <IonCardContent>
-                <IonText color="dark">
-                  <p>
-                    <b>Requires:</b> at least two experiments from the same
-                    dataset
-                  </p>
-                </IonText>
-              </IonCardContent>
-              <IonButton expand="full" fill="clear">
-                Start Benchmark
-              </IonButton>
-            </IonCard>
-          </IonCol>
-          <IonCol size="4" sizeXl="3">
-            <IonCard>
-              <IonCardHeader>
-                <IonCardTitle>
-                  N-Metrics Table
-                  <span>
-                    <IonIcon
-                      className="ion-float-right"
-                      icon={calculator}
-                      size="large"
-                      color="primarydark"
-                    />
-                  </span>
-                </IonCardTitle>
-              </IonCardHeader>
-              <IonCardContent>
+              </IonText>
+            </IonCardContent>
+            <IonButton expand="full" fill="clear">
+              Start Benchmark
+            </IonButton>
+          </IonCard>
+        </IonCol>
+        <IonCol size="4" sizeXl="3">
+          <IonCard>
+            <IonCardHeader>
+              <IonCardTitle>
+                N-Metrics Table
+                <span>
+                  <IonIcon
+                    className="ion-float-right"
+                    icon={calculator}
+                    size="large"
+                    color="primarydark"
+                  />
+                </span>
+              </IonCardTitle>
+            </IonCardHeader>
+            <IonCardContent>
+              <p>
+                Compare multiple experiments and a single gold standard.
+                You&apos;ll be able to compare all experiments&apos; metrics
+                against each other in a table.
+              </p>
+            </IonCardContent>
+            <IonCardContent>
+              <IonText color="dark">
                 <p>
-                  Compare multiple experiments and a single gold standard.
-                  You&apos;ll be able to compare all experiments&apos; metrics
-                  against each other in a table.
+                  <b>Requires:</b> at least two experiments from the same
+                  dataset
                 </p>
-              </IonCardContent>
-              <IonCardContent>
-                <IonText color="dark">
-                  <p>
-                    <b>Requires:</b> at least two experiments from the same
-                    dataset
-                  </p>
-                </IonText>
-              </IonCardContent>
-              <IonButton expand="full" fill="clear">
-                Start Benchmark
-              </IonButton>
-            </IonCard>
-          </IonCol>
-          <IonCol size="4" sizeXl="3">
-            <IonCard>
-              <IonCardHeader>
-                <IonCardTitle>
-                  Binary Metrics View
-                  <span>
-                    <IonIcon
-                      className="ion-float-right"
-                      icon={pauseCircle}
-                      size="large"
-                      color="primarydark"
-                    />
-                  </span>
-                </IonCardTitle>
-              </IonCardHeader>
-              <IonCardContent>
+              </IonText>
+            </IonCardContent>
+            <IonButton expand="full" fill="clear">
+              Start Benchmark
+            </IonButton>
+          </IonCard>
+        </IonCol>
+        <IonCol size="4" sizeXl="3">
+          <IonCard>
+            <IonCardHeader>
+              <IonCardTitle>
+                Binary Metrics View
+                <span>
+                  <IonIcon
+                    className="ion-float-right"
+                    icon={pauseCircle}
+                    size="large"
+                    color="primarydark"
+                  />
+                </span>
+              </IonCardTitle>
+            </IonCardHeader>
+            <IonCardContent>
+              <p>
+                Summary view to compare a single experiment against a ground
+                truth. You can view metrics as well as investigate tuples.
+              </p>
+            </IonCardContent>
+            <IonCardContent>
+              <IonText color="dark">
                 <p>
-                  Summary view to compare a single experiment against a ground
-                  truth. You can view metrics as well as investigate tuples.
+                  <b>Requires:</b> exactly two experiments from the same dataset
                 </p>
-              </IonCardContent>
-              <IonCardContent>
-                <IonText color="dark">
-                  <p>
-                    <b>Requires:</b> exactly two experiments from the same
-                    dataset
-                  </p>
-                </IonText>
-              </IonCardContent>
-              <IonButton expand="full" fill="clear">
-                Start Benchmark
-              </IonButton>
-            </IonCard>
-          </IonCol>
-          <IonCol size="4" sizeXl="3">
-            <IonCard>
-              <IonCardHeader>
-                <IonCardTitle>
-                  Diagram Investigator
-                  <span>
-                    <IonIcon
-                      className="ion-float-right"
-                      icon={barChart}
-                      size="large"
-                      color="primarydark"
-                    />
-                  </span>
-                </IonCardTitle>
-              </IonCardHeader>
-              <IonCardContent>
+              </IonText>
+            </IonCardContent>
+            <IonButton expand="full" fill="clear">
+              Start Benchmark
+            </IonButton>
+          </IonCard>
+        </IonCol>
+        <IonCol size="4" sizeXl="3">
+          <IonCard>
+            <IonCardHeader>
+              <IonCardTitle>
+                Diagram Investigator
+                <span>
+                  <IonIcon
+                    className="ion-float-right"
+                    icon={barChart}
+                    size="large"
+                    color="primarydark"
+                  />
+                </span>
+              </IonCardTitle>
+            </IonCardHeader>
+            <IonCardContent>
+              <p>
+                Investigate soft and hard KPIs of experiments grouped by either
+                dataset or matching solutions. Evaluation occurs with the help
+                of a variety of diagrams.
+              </p>
+            </IonCardContent>
+            <IonCardContent>
+              <IonText color="dark">
                 <p>
-                  Investigate soft and hard KPIs of experiments grouped by
-                  either dataset or matching solutions. Evaluation occurs with
-                  the help of a variety of diagrams.
+                  <b>Requires:</b> multiple experiments from any dataset
                 </p>
-              </IonCardContent>
-              <IonCardContent>
-                <IonText color="dark">
-                  <p>
-                    <b>Requires:</b> multiple experiments from any dataset
-                  </p>
-                </IonText>
-              </IonCardContent>
-              <IonButton expand="full" fill="clear">
-                Start Benchmark
-              </IonButton>
-            </IonCard>
-          </IonCol>
-        </IonRow>
-      </IonGrid>
-    </GenericSubApp>
+              </IonText>
+            </IonCardContent>
+            <IonButton expand="full" fill="clear">
+              Start Benchmark
+            </IonButton>
+          </IonCard>
+        </IonCol>
+      </IonRow>
+    </IonGrid>
   );
 };
