+++ app/src/pages/BenchmarkPage/BenchmarkApp.tsx
@@ -1,23 +1,21 @@
-import BenchmarkAppView from 'pages/BenchmarkPage/BenchmarkApp.View';
-import { BenchmarkAppDispatchProps } from 'pages/BenchmarkPage/BenchmarkAppProps';
-import { BenchmarkAppDispatch } from 'pages/BenchmarkPage/store/BenchmarkAppModel';
-import {
-  getAlgorithms,
-  getDatasets,
-  getExperiments,
-} from 'pages/BenchmarkPage/store/BenchmarkAppThunkActions';
-import { connect } from 'react-redux';
+import { SnowmanPublicState } from 'app/SnowmanPublicState';
+import GenericSubApp from 'components/GenericSubApp/GenericSubApp';
+import BenchmarkAppContainer from 'pages/BenchmarkPage/BenchmarkApp.Container';
+import BenchmarkSelector from 'pages/BenchmarkPage/components/BenchmarkSelector/BenchmarkSelector';
+import { createBenchmarkAppStore } from 'pages/BenchmarkPage/store/BenchmarkAppStoreFactory';
+import React from 'react';
+import { ViewIDs } from 'types/ViewIDs';
 
-const mapDispatchToProps = (
-  dispatch: BenchmarkAppDispatch
-): BenchmarkAppDispatchProps => ({
-  loadInitialState: () => {
-    dispatch(getAlgorithms()).then();
-    dispatch(getDatasets()).then();
-    //dispatch(getExperiments()).then();
-  },
-});
-
-const BenchmarkApp = connect(null, mapDispatchToProps)(BenchmarkAppView);
+const BenchmarkApp = (props: SnowmanPublicState): JSX.Element => (
+  <GenericSubApp
+    {...props}
+    appId={ViewIDs.DASHBOARD}
+    appTitle={'Benchmark Dashboard'}
+    createSubAppStore={createBenchmarkAppStore}
+    sideMenu={BenchmarkSelector}
+  >
+    <BenchmarkAppContainer />
+  </GenericSubApp>
+);
 
 export default BenchmarkApp;
